Refactor shape out into its own module
